requirements.txt
trino
pandas
loguru
fastapi
uvicorn[standard]
pydantic
langchain-openai
python-dotenv
sqlalchemy